backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-dotenv==1.0.0
requests==2.29.0
pydantic==1.10.7
sqlalchemy==2.0.12
aiohttp==3.8.4
pytest==7.3.1
httpx==0.24.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
redis==4.5.5
